+++ CHANGELOG.md
@@ -62,6 +62,7 @@ The new `price_via_btc` config option may not exist in your config file. If it i
 - KuCoin parser: added "Convert Orders_Filled Orders" from bundle files.
 - Conversion tool: added parser for Blockchair explorer. ([#490](https://github.com/BittyTax/BittyTax/issues/490))
 - Kraken parser: added "delistingconversion" and "airdrop" subtypes.
+- Koinly parser: added "Bulk edit in Excel" transactions export.
 ### Changed
 - Config: fiat_income to True.
 - Price tool: CoinDesk API deprecated.